2009-03-30  Aaron Bockover  <abockover@novell.com>
    * src/Mono.Zeroconf.Providers.Bonjour/Mono.Zeroconf.Providers.Bonjour/Service.cs:
    * src/Mono.Zeroconf.Providers.Bonjour/Mono.Zeroconf.Providers.Bonjour/BrowseService.cs:
    * src/Mono.Zeroconf.Providers.Bonjour/Mono.Zeroconf.Providers.Bonjour/RegisterService.cs:
    * src/Mono.Zeroconf/Mono.Zeroconf/IRegisterService.cs:
    * src/Mono.Zeroconf/Mono.Zeroconf/RegisterService.cs:
    * src/Mono.Zeroconf.Providers.AvahiDBus/Mono.Zeroconf.Providers.AvahiDBus/RegisterService.cs:
    Use ushort internally for the port value, add new UPort property on 
    IRegisterService so an unchecked cast is not necessary for port values that
    overflow Int16 (http://mono-project.com/Mono.Zeroconf#Known_Workarounds)


svn path=/trunk/Mono.Zeroconf/; revision=130586

+++ src/Mono.Zeroconf.Providers.Bonjour/Mono.Zeroconf.Providers.Bonjour/BrowseService.cs
@@ -113,7 +113,7 @@ namespace Mono.Zeroconf.Providers.Bonjour
             
             InterfaceIndex = interfaceIndex;
             FullName = fullname;
-            this.port = (short)port;
+            this.port = port;
             TxtRecord = new TxtRecord(txtLen, txtRecord);
 
             sdRef.Deallocate();
